FEAT: add update bmi when change weight, fix model

+++ health_infos/models.py
@@ -8,4 +8,4 @@ class Heath_Info(models.Model):
     height = models.DecimalField(max_digits=3, decimal_places=2)
     weight = models.DecimalField(max_digits=8, decimal_places=2)
     bmi = models.DecimalField(max_digits=4, decimal_places=2)
-    ideal_weight = models.DecimalField(max_digits=8, decimal_places=2)
+    ideal_weight = models.IntegerField()
